ready for PR to main, see desc.
DEV_NOTE: this is not final, but easily extendable utility;
